editeng: use EditLine as ref. and use const
We assert the EditLine pointer is valid (not nullptr) in some
methods or just expect it to be valid without asserting.  If this
is the case then we might as well change the parameter to be a
reference.

Also change EditLine to const&, where this makes sense in the code.

Change-Id: Ia28810fea9d4e053f18fbb45968644545f5bf83b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/162012
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>