CppunitTest_sc_tiledrendering: ignore invalid json payload for STATE_CHANGED
The interesting case for this testsuite is when the json for
LOK_CALLBACK_STATE_CHANGED has a commandName key, ignore it callback
when that's not the case.

See
<https://gerrit.libreoffice.org/c/core/+/161691/4#message-73673a8cedce0a5f661839553dd463983d22250a>
for a case where this can happen, the json payload was

"{\n    \"state\": \"0\"\n}\n"

Which can happen when SfxLokHelper::sendUnoStatus() thinks that the
commandName value would be empty.

Change-Id: I398319f33288742b8f46f0d07b112fd61e9a9731
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/161962
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins