tdf#159158 writerfilter: duplicate relativeHeights? last one wins
[When MS Word 2019 round-trips a file with a duplicate relativeHeight
 it replaces them with unique relativeHeights in the export.]

My THEORY is that bOldStyle should always be true for GraphicImport,
so a follow-up commit will basically revert all of this,
but at least if my theory is wrong, I'll still have fixed this problem.

Notice that there are two things that feed into zOrder,
the other being z-Index. I assume the very explicit code that says
that (except for bOldStyle) first one wins must be the standard for z-Index.

make CppunitTest_sw_ooxmlexport18 \
    CPPUNIT_TEST_NAME=testTdf159158_zOrder_duplicate

Change-Id: I842d016a7922427ce1290a7134d4e19fed52880f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/161989
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>