tdf#159158 writerfilter: relativeHeights of 0/1 are highest zOrder
Apparently 0 and 1 are the highest zOrders, 2 is the lowest, then 3...
At least, this seems to be true for relativeHeight.
Of course the documentation doesn't say anything about this,
but documents conclusively prove a maximum zOrder of 251,658,240.

[When MS Word 2019 round-trips a file with a relativeHeight of 0/1,
 it replaces them with the highest relativeHeight in the document,
 but values of 2+ are preserved.]

Notice that there are two things that feed into zOrder,
the other being z-index where negatives have special meaning
in this context, but no special meaning for 0 or 1.

Interestingly, relativeHeight is unsigned, but z-index is signed.
This ends up not being a problem with a maximum of 0F00 0000.

GOOD! LO normally exports starting from 2, but writes 0 if no zOrder.
// It seems 0 and 1 have special meaning:
// just start counting from 2 to avoid issues with that.

Proof case that 0 is higher than both 300 and 225,763,766
make CppunitTest_sw_ooxmlexport8 CPPUNIT_TEST_NAME=testN747461

make CppunitTest_sw_ooxmlexport18 \
    CPPUNIT_TEST_NAME=testTdf159158_zOrder_1and0equalA
make CppunitTest_sw_ooxmlexport18 \
    CPPUNIT_TEST_NAME=testTdf159158_zOrder_1and0equalB
make CppunitTest_sw_ooxmlexport18 \
    CPPUNIT_TEST_NAME=testTdf159158_zOrder_1and0max

Change-Id: I3540252ca5b0be3b2d847f1a3885adb850f39c39
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/162038
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>