tdf#159158 writerfilter: duplicate relativeHeights? last one wins #2
Revert the previous solution and replace it with this simpler one.

This version of the patch is riskier, but from what I can see
last one always wins for GraphicImport.
I assume that Miklos just limited it to a certain situation
as a sanity check to try to avoid causing regressions
outside of the scope he was working on.

The fact that all DOCX and RTF unit tests pass is a good sign.

The only way I would expect this to cause problems is
if z-index somehow can come into play with a GraphicImport
(since both z-index and relativeHeight feed into zOrder).
However, an assert didn't turn up any instances where
applyZOrder was called without a relativeHeight being set.
assert(m_rDomainMapper.IsRTFImport() || bOldStyle || UsedRelativeHeight);

Change-Id: I28e8acd5997eed0e82d5853c85b672c62b21afe5
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/162036
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>