android: Update gradle + wrapper to 8.5
Commit the result of running

    ./gradlew wrapper --distribution-type all --gradle-version 8.5 --gradle-distribution-sha256-sum=c16d517b50dd28b3f5838f0e844b7520b8f1eb610f2f29de7e4e04a1b7c9c79b

(minus the automatically added `networkTimeout=10000` in
`android/source/gradle/wrapper/gradle-wrapper.properties`).

This includes the following upstream change for gradle
to port from the use of `which` to the more portable
`command -v`: [1]

Thanks to Eli Schwartz for pointing this out in his previous
Gerrit change [2] that was suggesting to only backport
that particular change to the wrapper script.

[1] https://github.com/gradle/gradle/pull/23961
[2] https://gerrit.libreoffice.org/c/core/+/160667

Change-Id: I1c230c75d2d8583852c59ff43ddfdf77daf5d5ec
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/162081
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>