tdf#159158 writerfilter: relativeHeight always under z-index
No matter what z-index I chose, it always ends up being above
a relativeHeight. Since a heaven z-index is 0+,
that means that all relativeHeight's need to be treated as negative.

IsInHeaderFooter is a complication, because these are ALWAYS treated
as if they are below the text, and yet behindDoc relativeHeights
still need to be below their counterparts.

While this algorithm could never cover all the situations
(it would require transitioning to a sal_Int64 to do that)
it should be fine for practical purposes.

Plus, there are some unknown complications surrounding this.
If I increase the behindDoc relativeHeight by 1 in
tdf159158_zOrder_behindDocB.docx
the yellow star actually moves on top of the blue star,
which totally debunks my theory.

RTF doesn't really use relativeHeight,
but if it did, these changes should apply anyway.

Change-Id: I62e1f9ae8fe6f99534a2a799692120508204c553
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/162234
Reviewed-by: Justin Luth <jluth@mail.com>
Tested-by: Justin Luth <jluth@mail.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>