tdf#154863 Fix crashes when moving images in a large document
The problem is that when moving images around the code tries
to do things with anchored text frames that have already gone
through destruction.  The fix is to check if those frames have
the destruction bit set before using them, thus avoiding the
crashes.

Also, there is an assert that was firing and removing it
seems to have no negative effect in interacting with the file.

Change-Id: I899171ef3b5113f479725b0421f469c36e40e26c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/162151
Reviewed-by: Matt K <mattkse@gmail.com>
Tested-by: Matt K <mattkse@gmail.com>