sw: layout: prevent bad page break in table with 0-height rows
The bugdoc contains a table with most rows having a hidden-paragraph
field in every paragraph, and no cell border or padding.

The SwTabFrame is initially split across 6 pages, which should all be
joined, as there are only 3 visible rows, but the last follow is never
joined.

This is because SwTabFrame::Join() of the next-to-last follow doesn't
invalidate the SwTabFrame, because Grow(0) doesn't invalidate, and
later formatting anything in the moved SwRowFrame doesn't invalidate
either as it is all height 0.

Try to fix this by guessing that the row will have height 0 (even if
it's not formatted yet).

Change-Id: Ic8246ac91c741f5a215f89dc159c44d7c7cf2ce1
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/162203
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>
(cherry picked from commit ab7893544dc6be6dc192dffefd57cd5ddd421c35)