tdf#159243 fix packaging of Noto fonts/fix Armenian UI on macOS
different packages must not share the same package name

Packages create corresponding packagename.filelist files that are used during
the installset creation process. When multiple packages share the same name,
it is random (depending on the order make decides to build the targets) which
package will actually be included in the installation set. The last package
will "win" and have its file stored in the filelist and thus will be part of
the installation set, all others will not be considered.

Change-Id: Ieffd2addf6d59916c0c1fe01c64866472ff23803
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/162206
Tested-by: Jenkins
Reviewed-by: Christian Lohmaier <lohmaier+LibreOffice@googlemail.com>