tdf#155917 sw: fix Repeat of Insert File with Fly
There is a fly frame anchored on the 1st node of the inserted file.

On Repeat, first SwUndoInsLayFormat::RepeatImpl() runs and inserts
another fly, adding 3 nodes before the body of the document, then
SwUndoInserts::RepeatImpl() creates SwPaM from its stored positions that
is off by 3 nodes and invalid (crossing cell boundaries).

The problem is that the way the Undo objects are created in
SwReader::Read() is wrong, and there is alreay a better way not to
create the SwUndoInsLayFormat manually from the outside but
automatically in SwUndoInserts::SetInsertRange(), which is already used
in Paste - why Insert File never used this has always been a mystery.

The only difference appears to be that SwReader::Read() also handled
at-page anchored flys which SetInsertRange() didn't, as these probably
can't be part of a Paste, but that is easy to add.

(regression from commit 28b77c89dfcafae82cf2a6d85731b643ff9290e5)

Change-Id: I19d7a4d668ae8bed0986162da6ec5b5ea3118091
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/162208
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>