tdf#159158 DOCX import: shape's last duplicate z-index wins
It looks like for DOCX a duplicate always comes out on top,
which is the exact opposite of what we have been doing so far.

The ability to switch between "old" and "new" style was
added for RTF's benefit, and cautiously expanded to
include when RTF and DOCX IMPORT_AS_DETECTED_INLINE.

This does not affect RTF, since AFAIK RTF cannot contain
VML-Z-ORDER properties.

Change-Id: I110e3db3fa504f1ae3a5bba5979b40f0c00a467e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/162261
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>