Related: tdf#159259 Drop obsolete exclusion of framePr inside sdt
Commit 73bd937420b9a99e1e35950e3f9dcbcfd874876d (n#780853 fix DOCX
import of w:sdtContent in table cell, 2012-09-25) had introduced a
restriction, where a framePr was ignored, when inside an sdt. It
was required to allow table import, which broke otherwise.

Since then, the situation has obviously changed; the test case for
the change, added in commit 7ea71eb8a28c4b41949299ff7d4b391486c90eea
(n#780853 testcase, 2012-09-25) passes after removal of the check.

This check disallows importing sdt in frame. This makes the bugdoc
of tdf#159259 to import the sdt in frame. The Lorem Ipsum paragraph
is still enclosed into the sdt - to be fixed in a follow-up.

Change-Id: I60fbb2dd6f753682217d86e4f8d315f1883c6e83
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/162258
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>