Fix sw_complex JUnitTest not running
Call to gb_JunitTest_add_classes was missing in the gbuild makefile
sw/JunitTest_sw_complex.mk, and now it sholud be fixed. Before that
the sw/qa/complex/indeterminateState/CheckIndeterminateState.java was
compiled, but it was not run due to the missing class name
(complex.indeterminateState.CheckIndeterminateState) from the java
command which was used to invoke the test.

The test can be run with:

$ make JunitTest_sw_complex

Change-Id: I85823336d7f3ad2fde47442b5d1c5d2f3003299b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/162263
Tested-by: Jenkins
Reviewed-by: Hossein <hossein@libreoffice.org>