tdf#159023 sw_redlinehide: fix layout frames copying table into footer
Tables in footer aren't supported - if a table is pasted into the
footer, SwNodes::CopyNodes() inserts for table/cell nodes the
SwPlaceholderNodes so that the relative distance between any copied
node is the same in the target, and once the other stuff like bookmarks
and anchored flys are copied, SwNodes::DelDummyNodes() removes the
SwPlaceholderNodes.

If SwNodes::CopyNodes() is called with bNewFrames=true, it handles
creating the frames well, but with bNewFrames=false and a subsequent
MakeFrames(), the InsertCnt_() simply quits once it sees the first
SwPlaceholderNode and so no frames are created for the nodes in the
table cells.

Rearrange DocumentContentOperationsManager::CopyWithFlyInFly() to move
the MakeFrames() down below DelDummyNodes(), which requires adding a
parameter to CopyFlyInFlyImpl() to prevent it from creating frames
(anchored objects will also be created by MakeFrames() later).

(regression from commit 166b5010b402a41b192b1659093a25acf9065fd9)

Change-Id: I0ac30bec6e6423bb7d12c0a29f33848a5b61cd98
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/162320
Tested-by: Michael Stahl <michael.stahl@allotropia.de>
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>